Fix edge cases in the ACL controller when things could fail and never recover
Since ACL tokens are used to determine if the reconcile process needs to do
anything, there is a bug when ACL tokens are deleted or updated but the Secrets
Manager secrets aren't. This is fixed by making the `Upsert` function only
insert a secret once and then no longer default it again in the `Delete`
function.  This means that after the initial insert, updating and deleting
tokens only requires a single operation and can't end up in a partially
inserted or deleted state.  The originally insert still takes two operations
because it requires inserting the AWS secret and then inserting the ACL token
into Consul. Interesting cases:
* The only failure case occurs if the AWS secret is successfully inserted and
  then the ACL token fails to be deleted. When this happens, a subsequent
  reconcile will see that the AWS secret has been inserted and then reuse that
  secret when setting the `AccessorID` and `SecretID` on the ACL token it inserts.
* If a delete fails to delete the ACL token and then then tasks for that
  service start again, the secret and token already be set correctly and nothing
  will need to be reconciled for the service to run.
* If an upsert fails to insert the ACL token and then all tasks for the service
  are stopped, the token will be deleted in the next reconcile.